refactor(memfs): share post-turn sync flow
Centralize post-turn MemFS sync reminder handling and rename memory write helpers to reflect commit-only semantics.
