+++ src/agent/memory-git.ts
@@ -29,6 +29,7 @@ import {
   getMemfsGitProxyRewriteConfig,
   getMemfsServerUrl,
 } from "@/backend/api/memfs-git-proxy";
+import { SYSTEM_REMINDER_CLOSE, SYSTEM_REMINDER_OPEN } from "@/constants";
 import { debugLog, debugWarn } from "@/utils/debug";
 import { getScopedMemoryFilesystemRoot } from "./memory-filesystem";
 
@@ -61,7 +62,7 @@ export interface MemoryCommitAuthor {
   authorEmail: string;
 }
 
-export interface CommitAndSyncMemoryWriteParams {
+export interface CommitMemoryWriteParams {
   memoryDir: string;
   pathspecs: string[];
   reason: string;
@@ -71,7 +72,7 @@ export interface CommitAndSyncMemoryWriteParams {
 
 export type MemoryWriteSyncMode = "remote" | "local";
 
-export interface CommitAndSyncMemoryWriteResult {
+export interface CommitMemoryWriteResult {
   committed: boolean;
   sha?: string;
 }
@@ -1319,12 +1320,9 @@ async function unstageMemoryPaths(
   }
 }
 
-export async function assertMemoryRepoReadyForWrite(
+export async function assertMemoryRepoCleanForWrite(
   memoryDir: string,
-  _agentId?: string,
-  options: { syncMode?: MemoryWriteSyncMode } = {},
 ): Promise<void> {
-  void options;
   const status = await runGit(memoryDir, ["status", "--porcelain"]);
   if (status.stdout.trim().length > 0) {
     throw new Error(
@@ -1333,9 +1331,9 @@ export async function assertMemoryRepoReadyForWrite(
   }
 }
 
-export async function commitAndSyncMemoryWrite(
-  params: CommitAndSyncMemoryWriteParams,
-): Promise<CommitAndSyncMemoryWriteResult> {
+export async function commitMemoryWrite(
+  params: CommitMemoryWriteParams,
+): Promise<CommitMemoryWriteResult> {
   const normalizedPathspecs = normalizePathspecs(params.pathspecs);
   if (normalizedPathspecs.length === 0) {
     return { committed: false };
@@ -1688,6 +1686,14 @@ export interface MemoryPostTurnSyncResult {
   localOnly: boolean;
 }
 
+export interface RunPostTurnMemorySyncParams {
+  agentId: string;
+  isEnabled?: (agentId: string) => boolean;
+  enqueueReminder?: (text: string) => void;
+  emitWarning?: (text: string) => void | Promise<void>;
+  debugLabel?: string;
+}
+
 /**
  * Check git status of the memory directory.
  * Used to decide whether to inject a sync reminder.
@@ -1948,6 +1954,89 @@ export async function syncPendingMemoryCommitsAfterTurn(
         localOnly,
       };
     }
+  }
+}
+
+export function formatMemoryPostTurnSyncReminder(
+  result: MemoryPostTurnSyncResult,
+): string | null {
+  if (
+    result.status === "clean" ||
+    result.status === "pushed" ||
+    result.status === "skipped"
+  ) {
+    return null;
+  }
+
+  if (result.status === "conflict") {
+    return `${SYSTEM_REMINDER_OPEN}
+MEMORY GIT CONFLICT: The memory repository needs manual conflict resolution.
+
+Memory directory: ${result.memoryDir}
+Status: ${result.summary}
+
+Resolve the merge/rebase conflicts in the memory repository, stage the resolved files, and complete the merge/rebase or create the needed commit. The harness will retry remote push after a future turn when the repo is clean.
+${SYSTEM_REMINDER_CLOSE}`;
+  }
+
+  if (result.status === "dirty") {
+    const action = result.localOnly
+      ? "Commit these memory changes locally"
+      : "Commit these memory changes";
+    return `${SYSTEM_REMINDER_OPEN}
+MEMORY COMMIT NEEDED: The memory repository has uncommitted changes.
+
+Memory directory: ${result.memoryDir}
+Status: ${result.summary}
+
+${action} when appropriate. Do not run \`git push\` for MemFS sync; the harness pushes clean committed memory changes automatically for remote MemFS agents after turns.
+${SYSTEM_REMINDER_CLOSE}`;
+  }
+
+  return `${SYSTEM_REMINDER_OPEN}
+MEMORY SYNC FAILED: The harness could not push pending memory commits.
+
+Memory directory: ${result.memoryDir}
+Status: ${result.summary}
+
+Inspect the memory repository and resolve any local git issue. The harness will retry remote push after a future turn when the repo is clean.
+${SYSTEM_REMINDER_CLOSE}`;
+}
+
+export async function runPostTurnMemorySync(
+  params: RunPostTurnMemorySyncParams,
+): Promise<void> {
+  const debugLabel = params.debugLabel ?? "Post-turn memory sync";
+
+  try {
+    if (params.isEnabled && !params.isEnabled(params.agentId)) {
+      return;
+    }
+  } catch (error) {
+    debugWarn(
+      "memfs-git",
+      `Skipping ${debugLabel} because MemFS settings are unavailable: ${
+        error instanceof Error ? error.message : String(error)
+      }`,
+    );
+    return;
+  }
+
+  try {
+    const syncResult = await syncPendingMemoryCommitsAfterTurn(params.agentId);
+    const syncReminder = formatMemoryPostTurnSyncReminder(syncResult);
+    if (!syncReminder) {
+      return;
+    }
+    params.enqueueReminder?.(syncReminder);
+    await params.emitWarning?.(syncReminder);
+  } catch (error) {
+    debugWarn(
+      "memfs-git",
+      `${debugLabel} failed: ${
+        error instanceof Error ? error.message : String(error)
+      }`,
+    );
   }
 }
