+++ src/headless.ts
@@ -35,7 +35,7 @@ import { setAgentContext, setConversationId } from "./agent/context";
 import { createAgent } from "./agent/create";
 import { handleListMessages } from "./agent/list-messages-handler";
 import { ISOLATED_BLOCK_LABELS } from "./agent/memory";
-import { syncPendingMemoryCommitsAfterTurn } from "./agent/memory-git";
+import { runPostTurnMemorySync } from "./agent/memory-git";
 import { getStreamToolContextId, sendMessageStream } from "./agent/message";
 import {
   getModelInfo,
@@ -124,7 +124,6 @@ import {
   buildSharedReminderParts,
   prependReminderPartsToContent,
 } from "./reminders/engine";
-import { formatMemoryPostTurnSyncReminder } from "./reminders/memory-git-sync";
 import {
   createSharedReminderState,
   enqueueMemoryGitSyncReminder,
@@ -251,33 +250,6 @@ async function reportStartupErrorAndExit(
   }
 
   return await flushAndExit(1);
-}
-
-async function runPostTurnMemorySync(params: {
-  agentId: string;
-  enqueueReminder?: (text: string) => void;
-  emitWarning?: (text: string) => void | Promise<void>;
-}): Promise<void> {
-  if (!settingsManager.isMemfsEnabled(params.agentId)) {
-    return;
-  }
-
-  try {
-    const syncResult = await syncPendingMemoryCommitsAfterTurn(params.agentId);
-    const syncReminder = formatMemoryPostTurnSyncReminder(syncResult);
-    if (!syncReminder) {
-      return;
-    }
-    params.enqueueReminder?.(syncReminder);
-    await params.emitWarning?.(syncReminder);
-  } catch (error) {
-    debugWarn(
-      "memfs-git",
-      `Post-turn headless memory sync failed: ${
-        error instanceof Error ? error.message : String(error)
-      }`,
-    );
-  }
 }
 
 export type BidirectionalQueuedInput = QueuedTurnInput<
@@ -3018,6 +2990,8 @@ ${SYSTEM_REMINDER_CLOSE}
 
   await runPostTurnMemorySync({
     agentId: agent.id,
+    isEnabled: (id) => settingsManager.isMemfsEnabled(id),
+    debugLabel: "Post-turn headless memory sync",
     emitWarning: (text) => {
       if (outputFormat !== "stream-json") {
         console.error(text);
@@ -4558,6 +4532,8 @@ async function runBidirectionalMode(
       } finally {
         await runPostTurnMemorySync({
           agentId: agent.id,
+          isEnabled: (id) => settingsManager.isMemfsEnabled(id),
+          debugLabel: "Post-turn headless memory sync",
           enqueueReminder: (text) => {
             enqueueMemoryGitSyncReminder(sharedReminderState, { text });
           },
